requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-js-asset==1.2.2
Pillow==8.1.1
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1

# The Google APIs Client Library for Python:
google-api-python-client==2.13.0
# The google-auth, google-auth-oauthlib, 
# and google-auth-httplib2 for user authorization.
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-auth-httplib2==0.1.0

# django-model-utils for handy utils
django-model-utils==4.1.1

django-crispy-forms